Added leaderboard to command handler.
