+++ index.js
@@ -59,130 +59,9 @@ client.on('message', async msg => {
   let commandfile = client.commands.get(cmd.slice(prefix.length));
   if (commandfile) commandfile.run(client, msg, args);
 
-
+  
 
 	if (
-		msg.content.toLowerCase().startsWith(config.discordPrefix + 'staff') &&
-		msg.guild.id === '333669866735271938'
-	) {
-		const staff = [];
-		msg.guild.members.forEach(m => {
-			const roles = m.roles.map(r => r.name.toLowerCase());
-			if (!(m.presence.status === 'online' || m.presence.status === 'dnd'))
-				return;
-			if (roles.includes('moderator') || roles.includes('junior moderator'))
-				staff.push(m);
-			if (m.id === '142730862948122625' || m.id === '250536623270264833')
-				staff.push(m);
-		});
-		if (staff.length < 1) return msg.reply('No online staff right now');
-		msg.channel.send(
-			`Online staff:\n\n${staff.map(s => s.toString()).join(', ')}`
-		);
-	} else if (
-		msg.content
-			.toLowerCase()
-			.startsWith(config.discordPrefix + 'leaderboard') ||
-		msg.content.toLowerCase().startsWith(config.discordPrefix + 'lb')
-	) {
-		if (!args[0])
-			return msg.channel.send(
-				`**Usage!** ${config.discordPrefix}leaderboard (xp|kills|wins|losses)`
-			);
-		if (args[0].toLowerCase() === 'xp' || args[0].toLowerCase() === 'level') {
-			let response = await fetch(config.apiUrl + '/mc/leaderboard/xp');
-			let leaderboard = await response.json();
-			let lbMsg = [];
-			var count = 0;
-			leaderboard.slice(0, 10).forEach(player => {
-				count++;
-				if (count !== 11) {
-					lbMsg.push(
-						`${getNumberEmoji(count)} **${player.name}** (level ${
-							player.level
-						})`
-					);
-				}
-			});
-			let embed = new MessageEmbed();
-			embed.setTitle(`Top 10 players on ${config.servername} (sorted by xp)`);
-			embed.setColor('RED');
-			embed.setDescription(lbMsg.join('\n'));
-			msg.channel.send({ embed });
-		} else if (
-			args[0].toLowerCase() === 'kills' ||
-			args[0].toLowerCase() === 'kill'
-		) {
-			let response = await fetch(config.apiUrl + '/mc/leaderboard/kills');
-			let leaderboard = await response.json();
-			let lbMsg = [];
-			var count = 0;
-			leaderboard.slice(0, 10).forEach(player => {
-				count++;
-				if (count !== 11) {
-					lbMsg.push(
-						`${getNumberEmoji(count)} **${player.name}** (${
-							player.kills
-						} kills)`
-					);
-				}
-			});
-			let embed = new MessageEmbed();
-			embed.setTitle(`Top 10 players on ${config.servername} (sorted by kills)`);
-			embed.setColor('RED');
-			embed.setDescription(lbMsg.join('\n'));
-			msg.channel.send({ embed });
-		} else if (
-			args[0].toLowerCase() === 'wins' ||
-			args[0].toLowerCase() === 'win'
-		) {
-			let response = await fetch(config.apiUrl + '/mc/leaderboard/wins');
-			let leaderboard = await response.json();
-			let lbMsg = [];
-			var count = 0;
-			leaderboard.slice(0, 10).forEach(player => {
-				count++;
-				if (count !== 11) {
-					lbMsg.push(
-						`${getNumberEmoji(count)} **${player.name}** (${player.wins} wins)`
-					);
-				}
-			});
-			let embed = new MessageEmbed();
-			embed.setTitle(`Top 10 players on ${config.servername} (sorted by wins)`);
-			embed.setColor('RED');
-			embed.setDescription(lbMsg.join('\n'));
-			msg.channel.send({ embed });
-		} else if (
-			args[0].toLowerCase() === 'losses' ||
-			args[0].toLowerCase() === 'loss' ||
-			args[0].toLowerCase() === 'lose'
-		) {
-			let response = await fetch(config.apiUrl + '/mc/leaderboard/losses');
-			let leaderboard = await response.json();
-			let lbMsg = [];
-			var count = 0;
-			leaderboard.slice(0, 10).forEach(player => {
-				count++;
-				if (count !== 11) {
-					lbMsg.push(
-						`${getNumberEmoji(count)} **${player.name}** (${
-							player.losses
-						} losses)`
-					);
-				}
-			});
-			let embed = new MessageEmbed();
-			embed.setTitle(`Top 10 players on ${config.servername} (sorted by losses)`);
-			embed.setColor('RED');
-			embed.setDescription(lbMsg.join('\n'));
-			msg.channel.send({ embed });
-		} else {
-			return msg.channel.send(
-				`**Usage!** ${config.discordPrefix}leaderboard (xp|kills|wins|losses)`
-			);
-		}
-	} else if (
 		msg.content.toLowerCase().startsWith(config.discordPrefix + 'help')
 	) {
 		let embed = new MessageEmbed();
